Landing Page
Created Landing Page
Split online compiler and main page in 2 different webpack projects
Improved Playground look & feel
Added favicons
Moved tstl compiler and Lua VM to web worker instances
Added build and debug scripts
