+++ assets/layout/play.html
@@ -1,0 +1,21 @@
+<div id="header">
+    <div id="logo">
+        <span class="logo-bold">Type</span>ScriptTo<span class="logo-bold">Lua</span> Playground
+    </div>
+</div>
+<div id="content">
+    <div id="editor-container">
+        <div id="editor-ts" class="editor">
+        </div>
+        <div id="editor-lua" class="editor">
+        </div>
+        <div id="editor-output" class="editor">
+            <div id="editor-output-linenumbers">
+                >_
+            </div>
+            <div id="editor-output-content">
+
+            </div>
+        </div>
+    </div>
+</div>
