+++ assets/styles/_base.scss
@@ -1,0 +1,15 @@
+@import 'fonts';
+@import 'colors';
+
+html, body {
+    width: 100%; height: 100%;
+    padding: 0px;
+    margin: 0px;
+    font-family: "Open Sans";
+    background-color: $body-background;
+    color: $body-font-color;
+}
+
+div {
+    box-sizing: border-box;
+}
