+++ assets/styles/_colors.scss
@@ -1,0 +1,6 @@
+$lua-purple: rgb(30, 40, 111);
+$ts-blue: rgb(67, 119, 194);
+$body-background: #1e1e1e;
+$grey-accent: #333333;
+$body-font-color: #ededed;
+$card-color: #1d2021;
