+++ assets/styles/_fonts.scss
@@ -1,0 +1,3 @@
+@import url('https://fonts.googleapis.com/css?family=Open+Sans:300,400,700|Raleway:300,400,700');
+
+$console-fonts: "Lucida Console", Monaco, monospace;
