+++ assets/styles/play.scss
@@ -1,0 +1,80 @@
+$header-height: 40px;
+$output-height: 120px;
+
+@import 'base';
+
+#header {
+    width: 100%;
+    height: $header-height;
+    //border-bottom-left-radius: 10%;
+    //border-bottom-right-radius: 60%;
+    //background: linear-gradient(135deg, $ts-blue, $lua-purple);
+    //-webkit-clip-path: polygon(100% 0, 100% 70%, 80% 90%, 50% 100%, 20% 90%, 0% 70%, 0 0);
+    //clip-path: polygon(100% 0, 100% 70%, 80% 90%, 50% 100%, 20% 90%, 0% 70%, 0 0);
+    background-color: $grey-accent;
+    box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
+    z-index: 10;
+}
+
+#logo {
+    position: absolute;
+    top: 7px;
+    left: 10px;
+    font-weight: 300;
+    font-size: 20px;
+}
+
+.logo-bold {
+    font-weight: 400;
+}
+
+#content {
+    height: calc(100% - #{$header-height});
+}
+
+#editor-container {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-wrap: wrap;
+
+    .editor {
+        height: calc(100% - #{$output-height});
+        background-color:$body-background;
+        color: white;
+        display: inline-block;
+        overflow: hidden;
+        border-bottom: 1px $grey-accent solid;
+    }
+
+    #editor-ts {
+        flex: 50%;
+    }
+
+    #editor-lua {
+        flex: 50%;
+    }
+    
+    #editor-output {
+        height: $output-height;
+        flex: 100%;
+        overflow: scroll;
+        font-family: $console-fonts;
+        display: flex;
+    }
+
+    #editor-output-linenumbers {
+        width: 40px;
+        height: 100%;
+        border-right: 1px $grey-accent solid;
+        padding-top: 10px;
+        padding-left: 10px;
+        padding-right: 10px;
+    }
+
+    #editor-output-content {
+        padding-top: 10px;
+        padding-left: 10px;
+        width: calc(100% - 20px);
+    }
+}
