+++ package.json
@@ -4,16 +4,22 @@
   "description": "",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack"
+    "build-playground": "webpack --config ./src/playground/webpack.config.js --mode production",
+    "dev-playground": "webpack-dev-server --content-base dist/ --config src/playground/webpack.config.js --mode development",
+    "build-landing": "webpack --config ./src/landing/webpack.config.js --mode production",
+    "dev-landing": "webpack-dev-server --content-base dist/ --config src/landing/webpack.config.js --mode development",
+    "build": "npm run build-playground && npm run build-landing"
   },
   "keywords": [],
   "author": "",
-  "license": "MIT",
+  "license": "",
   "devDependencies": {
+    "@types/highlight.js": "^9.12.3",
     "@types/node": "^10.11.7",
     "css-loader": "^1.0.0",
     "file-loader": "^2.0.0",
     "gh-pages": "^2.0.1",
+    "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "monaco-editor-webpack-plugin": "^1.5.4",
     "node-sass": "^4.9.3",
@@ -26,11 +32,12 @@
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.9",
+    "webpack-merge": "^4.1.4",
     "worker-loader": "^2.0.0"
   },
   "dependencies": {
-    "@types/url-parse": "^1.4.1",
     "fengari-web": "^0.1.2",
+    "highlight.js": "^9.13.1",
     "monaco-editor": "^0.14.3",
     "typescript-to-lua": "^0.10.0"
   }
