--- src/assets/layout/index.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="apple-touch-icon" sizes="180x180" href="<%= require('../images/favicons/apple-touch-icon.png') %>">
-    <link rel="icon" type="image/png" sizes="32x32" href="<%= require('../images/favicons/favicon-32x32.png') %>">
-    <link rel="icon" type="image/png" sizes="16x16" href="<%= require('../images/favicons/favicon-16x16.png') %>">
-    <link rel="manifest" href="<%= require('../images/favicons/site.webmanifest') %>">
-    <link rel="mask-icon" href="<%= require('../images/favicons/safari-pinned-tab.svg') %>" color="#418fd6">
-    <meta name="msapplication-TileColor" content="#2d89ef">
-    <meta name="theme-color" content="#ffffff">
-    <title>TypeScriptToLua - Online Compiler</title>
-</head>
-<body>
-    <div id="header">
-        <div id="logo">
-            <img class="logo-element" id="logo-ts" src="<%= require('../images/typescript_logo.svg') %>"/>
-            <div class="logo-element" id="logo-spacer">{to}</div>
-            <img class="logo-element" id="logo-lua" src="<%= require('../images/lua_logo.svg') %>"/>
-        </div>
-        <!--<div id="headline">
-            A Typescript to Lua Transpiler.
-        </div>-->
-    </div>
-    <div id="content">
-        <div id="example">
-            <div id="example-ts" class="example-element">
-            </div>
-            <div id="example-lua" class="example-element">
-            </div>
-            <!--<div id="example-control" class="example-element">
-                RUN
-            </div>-->
-            <div id="example-output" class="example-element">
-                Output
-            </div>
-        </div>
-    </div>
-    
-</body>
-</html>
