--- src/assets/styles/_fonts.scss
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Open+Sans:400,700|Raleway:400,700');
