--- src/assets/styles/main.scss
@@ -1,116 +1,0 @@
-$lua-purple: rgb(30, 40, 111);
-$ts-blue: rgb(67, 119, 194);
-$body-background: #1e1e1e;
-
-@import 'fonts';
-
-html, body {
-    width: 100%; height: 100%;
-    padding: 0px;
-    margin: 0px;
-    font-family: "Open Sans";
-    background-color: $body-background;
-}
-
-img {
-    width: 100%;
-    height: auto;
-}
-
-#header {
-    width: 100%;
-    height: 55px;
-    //border-bottom-left-radius: 10%;
-    //border-bottom-right-radius: 60%;
-    background: linear-gradient(135deg, $ts-blue, $lua-purple);
-    //-webkit-clip-path: polygon(100% 0, 100% 70%, 80% 90%, 50% 100%, 20% 90%, 0% 70%, 0 0);
-    //clip-path: polygon(100% 0, 100% 70%, 80% 90%, 50% 100%, 20% 90%, 0% 70%, 0 0);
-    color: white;
-    box-shadow: 0px 3px 5px 0px rgba(0,0,0,0.75);
-}
-
-#logo {
-    position: absolute;
-    top: 5px;
-    left: 5px;
-
-    .logo-element {
-        display: inline-block;
-        width: auto;
-        vertical-align: middle;
-    }
-
-    #logo-ts {
-        height: 30px;
-    }
-
-    #logo-spacer {
-        font-size: 30px;
-        vertical-align: -6px;
-    }
-
-    #logo-lua {
-        height: 40px;
-    }
-}
-
-#headline {
-    font-size: 48px;
-    font-weight: bold;
-    padding: 100px 150px 0 10%;
-    width: 750px;
-    position: relative;
-    float: right;
-}
-
-#headline:after {
-    content: "";
-    display: block;
-    margin-top: 10px;
-    border-bottom: solid 1px rgba(240, 240, 240, 0.5);
-}
-
-#content {
-}
-
-
-#example {
-    display: flex;
-    flex-wrap: wrap;
-    width: 100%;
-
-    .example-element {
-        flex: 1;
-        height: 600px;
-        background-color:$body-background;
-        display: inline-block;
-        color: white;
-        overflow: hidden;
-        border-bottom: 1px #222222 solid;
-        border-right: 1px #222222 solid;
-        border-left: 1px #222222 solid;
-    }
-
-    #example-ts {
-    }
-
-    #example-lua {
-    }
-
-    #example-controls {
-        height: 40px;
-        margin-top: 20px;
-        flex-basis: 100%;
-    }
-    
-    #example-output {
-        height: 120px;
-        flex-basis: 100%;
-    }
-}
-
-.center-block {
-    display: block;
-    margin-right: auto;
-    margin-left: auto;
-}
