+++ src/playground/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+      "sourceMap": true,
+      "noImplicitAny": true,
+      "module": "commonjs",
+      "target": "es5",
+      "lib": ["es5", "scripthost", "dom", "webworker"],
+      "allowJs": true,
+      "strict": true,
+    },
+    "exclude": [
+      "webpack*.js"
+    ]
+  }
